Cosmetic tweak of Unit tests
