+++ Tests/Unit/EventStore/Normalizer/Fixture/ArrayBasedValueObject.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\EventSourcing\Tests\Unit\EventStore\Normalizer\Fixture;
+
+class ArrayBasedValueObject
+{
+    public static function fromArray(array $value): self
+    {
+        return new self();
+    }
+}
