+++ Tests/Unit/EventStore/Normalizer/Fixture/IntegerBasedValueObject.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Neos\EventSourcing\Tests\Unit\EventStore\Normalizer\Fixture;
+
+class IntegerBasedValueObject
+{
+    public static function fromInteger(int $value): self
+    {
+        return new self();
+    }
+}
